Add force_unlock/powermeter/telemetry to evsemanager
The force_unlock command should be used by OCPP if requested by central
management system.

The powermeter interface is proxied to the actual power meter in use by
the evse to charge the car. Connect the OCPP powermeter requirement to
this one instead of directly connecting it to the actual hw driver.

A telemetry var is now being published that contains more data such as
rpm and health data.

Signed-off-by: Cornelius Claussen <cc@pionix.de>