Fix issue with multiple service list items
The mechanism for adding new service list items has been reworked. Now
there is a separate function which is responsible to add new services to
the service list. Now it is no longer possible to add multiple service
list items of the same service ID to the service list. In case the
function is called with a service ID which has been already configured,
the service list item will now be overwritten.

Signed-off-by: Fabian Hartung <fabian.hartung@chargebyte.com>