Allow Modules in sub directories
everest-core has many different modules now and it is hard to find out
the type of a module, e.g. if it is a hardware driver or a simulation
module etc.
This PR allows to move modules into subfolders in the source tree. Upon
install, the modules will still be installed flat into one folder. Two
modules are moved to a deprected folder as an example. All other modules
are not moved as this would create a lot of conflicts with other
branches. Moving into a new structure needs to be a coordinated effort,
but this adds the basis to be able to do that.

Signed-off-by: Cornelius Claussen <cc@pionix.de>