Move generated operator<< for types into namespace
We stumbled over the problem that we cannot pass types directly to
EVLOG_*. According to:
https://stackoverflow.com/questions/74451640/overloading-ostream-operator-works-for-stdcout-but-not-for-boostlog

Quote:
Your operator<< must be able to be found using ADL from Boost.Log namespace.
For this, the operator must be defined in the namespace of one of its argument types.

So let's move the generated operator a few line highe into the namespace.

Signed-off-by: Michael Heimpold <michael.heimpold@chargebyte.com>