Allow setting user for all modules and capabilities per module (#100)
* Allow setting user for all modules and capabilities per module

When setting run_as_user="myuser" property under settings in a config file
all modules are started as this user (along its gid and all other groups it belongs to)

After starting all modules the manager also switches to this user if the controller is disabled (which is used by the admin panel). If the controller is compiled in the manager will continue to run as root as it needs
to restart the modules if requested through the controller.
On production deployments the controller should always be disabled.

For each module, special Linux capabilities can be specified with e.g.
capabilities: "CAP_NET_RAW" in the module section. Multiple capabilities can 
be specified.

The manager will need to start as root or sufficient capabilities tobe able to change the user or set capabilities.
If no user/capabilities are specified all modules and the manager will not change the user during runtime.

A typical scenario is that the manager is started as root from systemd and then changes to a unprivileged user, while a few modules will get required capabilities to e.g. perform SLAC (CAP_NET_RAW).

This commit also increases everest-framework version number to 0.9

Signed-off-by: aw <aw@pionix.de>
Signed-off-by: Kai-Uwe Hermann <kai-uwe.hermann@pionix.de>
Signed-off-by: Cornelius Claussen <cc@pionix.de>