SmartChargingHandler: Add validation for TxProfiles (#452)
* ComponentStateManager: Create interface for testability

Splits the ComponentStateManager API out into an interface so
that it's easier to write unit tests that make use of its functionality.

In `transaction_metervalues_insert ()`, we were accessing the
first element of a list, then checking whether or not it was
empty. This causes an exception since if this list *is* empty,
there's no element to access.

This commit shifts the code around so that the empty check occurs
before accessing any elements of the list.

* SmartChargingHandler: Add validation for TxProfiles

Creates a skeleton for SmartChargingHandler and adds a method
to validate TxProfiles and TxProfiles only. This method
ensures that a given profile (assumed to be TxProfile) fits all of
the functional requirements specific to TxProfiles per K01.

The function, `validate_tx_profile ()`, takes a profile and an
EVSE, and returns a variant of `ProfileValidationResultEnum`.
This enum allows us to have detailed responses for the different types
of invalid profiles, and eventually will let us log *why* a profile
is invalid if the general validation fails.

Part of the prerequisites for https://github.com/EVerest/libocpp/issues/361

---------

Signed-off-by: Christopher Davis <150722105+christopher-davis-afs@users.noreply.github.com>
Co-authored-by: Gianfranco Berardi <54074967+gberardi-pillar@users.noreply.github.com>