feat: create YAML/JSON definition for charging schedule
Note this is an API breaking change. Originally charging schedules were
published:

{ "0": {...}, "1": {...}}

using the connector number as the key. This is not supported
in the YAML interface or type definitions.
Instead an array is used with an additional key added:

[ { "connector": 0, ...}, { "connector": 1, ...} ... ]

Note is should not be assumed that the array index will match the connector
number.

The existing types in libocpp have not been changed. A new mapping
has been created in the OCPP module.
(OCPP201 doesn't publish schedules at the moment)

Signed-off-by: James Chapman <james.chapman@pionix.de>