feat: added thread safe event queue and enum based flags
Moved ActiveErrors to be based on bit flags so that checking for
all_cleared() is no longer prone to missing an item in the check.

An event queue is used to replace the use of threads to update
shared_context. The queue ensures that the order of events is
maintained.

Previously signal_error.connect thread and signal_all_errors_cleared.connect
thread could both be blocked on state_machine_mutex and it is not defined
which thread would be active first.

Signed-off-by: James Chapman <james.chapman@pionix.de>