Added charging schedules definition (#582)
* feat: create YAML/JSON definition for charging schedule

Note this is an API breaking change. Originally charging schedules were
published:

{ "0": {...}, "1": {...}}

using the connector number as the key. This is not supported
in the YAML interface or type definitions.
Instead an array is used with an additional key added:

[ { "connector": 0, ...}, { "connector": 1, ...} ... ]

Note is should not be assumed that the array index will match the connector
number.

The existing types in libocpp have not been changed. A new mapping
has been created in the OCPP module.
(OCPP201 doesn't publish schedules at the moment)

Signed-off-by: James Chapman <james.chapman@pionix.de>

* fix: Arrays not supported by the autogenerated code

The autogenerated code doesn't support a module var being an array
(of a specific object type)

[{},{}]

Instead the array is now wrapped in an object:

{"schedules": [{},{}]}

Signed-off-by: James Chapman <james.chapman@pionix.de>

* feat: add optional EVSE id for OCPP 2.0.1

Signed-off-by: James Chapman <james.chapman@pionix.de>

* fix: addressed PR comments on use of snake_case
fix: moved conversions to separate file

Signed-off-by: James Chapman <james.chapman@pionix.de>

---------

Signed-off-by: James Chapman <james.chapman@pionix.de>
Co-authored-by: Piet Gömpel <37657534+Pietfried@users.noreply.github.com>