fix: address mutex deadlock in state machine
There was a deadlock in the IEC state machine when the C1 timer
expires whilst the state machine is running and attempting to
stop the timer.

The timer riases an event (hence blocks on the mutex)
Stopping the timer waits for the thread to finish.
Since the state machine is holding the mutex the thread
never finishes ...

fix: improvement to EventQueue

Wait on the condition variable now checks the queue size to
avoid spurious wakeups.
https://en.cppreference.com/w/cpp/thread/condition_variable

Signed-off-by: James Chapman <james.chapman@pionix.de>