Refactor power supply dc interface
This new interface is easier to use. It changes the way the capabilities
are used:

The sync get_capabilities command was removed in favour of async
publishing of capabilities from the driver module.

Now the driver module can publish its own capabilities as soon as they
are known (should be done asap after start up) and any later point in
time to update them.

It now also uses the error framework to report errors.

Signed-off-by: Cornelius Claussen <cc@pionix.de>