System: introduce ResetDelay config option
This configuration key allows to delay the execution of the reset
by a small amount of time.

A use-case could be, for example, when the actual request to reset
is received via the network and the caller wants to receive a
confirmation that the request will be executed. Performing the
reset directly in this case also immediately closes every
communication channel so that the caller never receive the confirmation.

This new delay allows the confirmation to travel back to the caller.

It is assumed that users configure reasonable values so there is no
enforced maximum or any other check for now.

Signed-off-by: Michael Heimpold <michael.heimpold@chargebyte.com>