Add source to enable/disable commands/events (#505)
* Add new enable_disable commands to EvseManager

It replaces the old enable/disable commands and adds a source and
priority. See README of the API module for more details.

Compile time breaking change on the evse_manager interface.
All enable/disable calls need to be switched over to the new
enable_disable call, for everest-core this is done in this PR.

In some uses cases it is useful if e.g. a LocalAPI source disables
a connector that only the same local source can enable again.
Think of a service technician who has an app to disable the connector
locally. Then it should be prevented that it is re-enabled by the CSMS
until the service technician enables it again via the same app.
Using priorities this is quite flexible and a lot of use-cases can
be implemented without a change in EVerest code.

The API module also has a new enable_disable command. It also has
compatibility implementations for the old enable/disable commands.
---------

Signed-off-by: James Chapman <james.chapman@pionix.de>
Signed-off-by: Piet Gömpel <37657534+Pietfried@users.noreply.github.com>
Signed-off-by: pietfried <pietgoempel@gmail.com>
Co-authored-by: James Chapman <james.chapman@pionix.de>
Co-authored-by: Piet Gömpel <37657534+Pietfried@users.noreply.github.com>
Co-authored-by: pietfried <pietgoempel@gmail.com>