Cleanup dangling transactions on startup
If EvseManager has a connection to a PersistentStore/kvs module, it now
stores the transaction UUID persistently.

It checks on startup if a transaction is still ongoing and tries to
close it in the power meter to fetch the OCMF data.
Then it emits a TransactionFinished event (with OCMF included if the
powermeter transaction stop was successful, otherwise the event is sent
without OCMF).

After that it clears any other transaction that still may be ongoing in
the powermeter by calling stop_transaction with an empty string as
session id.

LEM driver was adapted to support the cleanup with empty UUIDs on start
up.

Signed-off-by: Cornelius Claussen <cc@pionix.de>
Signed-off-by: florinmihut <florinmihut1@gmail.com>
Signed-off-by: pietfried <pietgoempel@gmail.com>