Make correct phase count available in API (#876)
The available phase count is part of the type "evse_manager::limits".
This type is published via mqtt and is used by the API module to interact
with external software stacks like Nymea.
Per default the available phase count is set to value 1, during the init
of the evse_manager interface. The EvseManager determines the count of
active phases with help of the min and max phase count values as part of
the type "evse_board_support::HardwareCapabilities". This is published by other
modules like CbTarragonDriver.
But this value is never published via "limits" topic through API module.
This leads to wrong values in the Nymea App. Although 3 phases are connected
to wallbox, only 1 phase is used for energy management.
This commit should fix this.

Signed-off-by: Martin Lukas <martin.lukas@chargebyte.com>