OCPP 2.0.1: Refactor Smart Charging Persistence (#790)
* smart_charging: Store ChargingLimitSource

When calling `validate_and_add_profile()`, `add_profile()`, and
`insert_or_update_charging_profile()`, we now can be given a source for
the profile. This source is stored in our database and can be retrieved
with the new method, `get_charging_limit_source_for_profile()`.

We use this new method to retrieve the source when retrieving profiles
for K09.

All new functionality is under test.

Co-authored-by: Coury Richards <146002925+couryrr-afs@users.noreply.github.com>
Signed-off-by: Christopher Davis <150722105+christopher-davis-afs@users.noreply.github.com>

* database_handler: Add function to get profiles by EVSE ID

Signed-off-by: Christopher Davis <150722105+christopher-davis-afs@users.noreply.github.com>

* smart_charging: Use new database handler func

Signed-off-by: Christopher Davis <150722105+christopher-davis-afs@users.noreply.github.com>

* database_handler: Add function to get all profiles in one vector

Signed-off-by: Christopher Davis <150722105+christopher-davis-afs@users.noreply.github.com>

* smart_charging: Only store and load profiles from db

Remove the map of profiles from the smart charging handler and use the
database for storing, loading, and finding profiles.

We've removed the function for getting all profiles from the
smart charging handler, and in all cases where that was used, including
tests, we retrieve the profiles from the database.

Signed-off-by: Christopher Davis <150722105+christopher-davis-afs@users.noreply.github.com>

* database_handler: Fix storage of charging profile purpose

SQLite needs to hold onto a copy of the string in order for it to be
stored properly.

Signed-off-by: Christopher Davis <150722105+christopher-davis-afs@users.noreply.github.com>

* database_handler: Store transaction ID on profiles

Edit the migration file for charging limit sources and remove

Signed-off-by: Christopher Davis <150722105+christopher-davis-afs@users.noreply.github.com>

* database_handler: Add function for new statements

Signed-off-by: Christopher Davis <150722105+christopher-davis-afs@users.noreply.github.com>

* smart_charging: Query database for matches directly

Use the database handler to select matching profiles instead of
iterating over the entire list of profiles.

Signed-off-by: Christopher Davis <150722105+christopher-davis-afs@users.noreply.github.com>

* smart_charging, database_handler: Use db for K10

Use new functions on the database handler to
reimplement clearing charging profiles by a given
set of criteria. This removes the existing
implementation within the smart charging handler.

Signed-off-by: Christopher Davis <150722105+christopher-davis-afs@users.noreply.github.com>

* smart_charging, types: Move ReportedChargingProfiles

Move the ReportedChargingProfiles struct to types.hpp where it can
be used in e.g. DatabaseHandler without causing recursive imports.

Signed-off-by: Christopher Davis <150722105+christopher-davis-afs@users.noreply.github.com>

* smart_charging, database_handler: Use db for K09

Use new functions on the database handler to
implement getting charging profiles by criteria.
This allows us to remove more places where we were
iterating over the entire list of charging
profiles.

This commit includes a fix for a bug where we
were only accepting searches if one of the sources
we were searching for is CSO.

Co-authored-by: Christopher Davis <150722105+christopher-davis-afs@users.noreply.github.com>
Signed-off-by: Coury Richards <146002925+couryrr-afs@users.noreply.github.com>
Signed-off-by: Christopher Davis <150722105+christopher-davis-afs@users.noreply.github.com>

* smart_charging: Use db for overlapping validity periods

Signed-off-by: Christopher Davis <150722105+christopher-davis-afs@users.noreply.github.com>

* dependencies: Rename json_schema_validator dep

The name wasn't quite right, leading to some confusion.

Signed-off-by: Christopher Davis <150722105+christopher-davis-afs@users.noreply.github.com>

* Renamed the method load charging profiles to clear invalid charging profiles, removed logic that adds profiles and replaced it with logic that only removes invalid profiles

Signed-off-by: Peter Giavotto <146003699+Giavotto@users.noreply.github.com>

---------

Signed-off-by: Christopher Davis <150722105+christopher-davis-afs@users.noreply.github.com>
Signed-off-by: Coury Richards <146002925+couryrr-afs@users.noreply.github.com>
Signed-off-by: Peter Giavotto <146003699+Giavotto@users.noreply.github.com>
Co-authored-by: Coury Richards <146002925+couryrr-afs@users.noreply.github.com>
Co-authored-by: Peter Giavotto <146003699+Giavotto@users.noreply.github.com>