Fix: MQTTAbstractionImpl: flush/sync MQTT Connect message immediately (#208)
* Fix: MQTTAbstractionImpl: flush/sync MQTT Connect message immediately

Previously, a socket was opened, but the MQTT Connect command message was
not flushed to the broker until init() was finished. This could be critical
if a module's init() phase waits for some reason, as the MQTT broker's
default socket timeout might disconnect the client.

This behavior was observed with the mosquitto broker, where unregistered
sockets were consistently disconnected after 94 seconds.

Insert an additional mqtt_sync() directly after mqtt_connect(), while the
mqtt_sync() thread is not yet started.

Co-authored-by: Fabian Hartung <fabian.hartung@chargebyte.com>
Signed-off-by: Moritz Barsnick <moritz.barsnick@chargebyte.com>

* Integrate early returns

This commit integrates early returns in the mqtt_abstraction_impl.cpp
file. This is done to improve the readability of the code and to reduce
the nesting of the code.

Signed-off-by: Fabian Hartung <fabian.hartung@chargebyte.com>

---------

Signed-off-by: Moritz Barsnick <moritz.barsnick@chargebyte.com>
Signed-off-by: Fabian Hartung <fabian.hartung@chargebyte.com>
Co-authored-by: Fabian Hartung <fabian.hartung@chargebyte.com>