EnergyManager: JSON based unit tests
Essentially input and expected output of the run_optimizer function can be specified in a single JSON file.
One unit test per file will programmatically be added to compare the output with the expected output and show a JSON diff if the test fails.

Several new test cases are added and the previous C++ based unit tests are converted to JSON.

Signed-off-by: Cornelius Claussen <cc@pionix.de>
Signed-off-by: Andreas Heinrich <andreas.heinrich@rwth-aachen.de>
Signed-off-by: Piet Gömpel <pietgoempel@gmail.com>
Signed-off-by: Abby Wheelis <Abigail.Wheelis@nrel.gov>