API: add support for optional module dependency 'charger_info'
The current implementation allows to load general charger
information from a YAML file which is then provided via MQTT.

The newer ChargerInfo module can provide the same information.
This change adapts the API module so that it can honour this.

To allow backward compatibility, the functionlity to load
the data from the file is still left as-is. However, it is not
used (the loaded data is replaced) when the optional
dependency is used/configured.

Note: The current/older keys in the API's YAML file differ from
the new ChargerInfo interface. This is kept as-is to ensure
backwards compatibility. But it is recommend to switch
to the newer interface to have a uniform presentation of
EVerest powered charging stations.

Signed-off-by: Michael Heimpold <mhei@heimpold.de>