Refactor interface & sdp server handling (#50)
Added an interface check. If the interface is auto, the first found ipv6 interface will be selected. If no prober interface is provided, the tbd_controller throws an exception. The port from the sdp server is now reusable. The sdp socket binds now to the interface and joins multicast group

Signed-off-by: Sebastian Lukas <sebastian.lukas@pionix.de>