feat(IsoExtension/EvseManager): Add vars to interfaces
We need some information about the EV to fill out the ConnectedEV
component, this is now possible thanks to this new var of the
iso15118_extensions interface.
We also added information whether service renegotiation is supported.

A new config was added to the evse manager to list the supported
operation modes. We also added vars to transmit what is the full list of
supported energy transfers and operation modes.

Signed-off-by: Martin Litre <mnlitre@gmail.com>