feat(OCPP2.1): Add AllowedEnergyTransferModes to auth validator (#1200)
* feat(OCPP2.1): Add AllowedEnergyTransferModes to auth validator

This list of allowed energy transfer modes has been added to the auth
validator interface so that the evse manager can be aware of what energy
transfer modes are allowed so that it can properly update the ISO15118
stack.

Signed-off-by: Martin Litre <mnlitre@gmail.com>

* feat: Add update energy transfer modes to both EvseManager/Iso15518

Added the possibility to update the supported energy transfer modes at
run time. This is necessary for use case Q01 for OCPP 2.1 where we want
to dynamically control whether or not we want to allow V2X.

Signed-off-by: Martin Litre <mnlitre@gmail.com>

---------

Signed-off-by: Martin Litre <mnlitre@gmail.com>