Fix json exceptions being thrown during handling of invalid messages (#1107)
* Fix json exceptions being thrown during handling of invalid messages

When certain malformed messages that during parsing result
in a json exceptions are reported to the CSMS it was observed
that secondary parsing exceptions can occur when the error message
is sent to the CSMS as part of a CallError as well as when the
incoming message in re-serialized in the security event to be sent.

This fixes #1104 by catching the secondary exception thrown during
error message generation, and by not sending a truncated version
of the problematic message but the error message instead.

* Handle json errors in 1.6 message_callback similarly to 2.x code

---------

Signed-off-by: Kai-Uwe Hermann <kai-uwe.hermann@pionix.de>