Feature/ocpp21 bidirectional building blocks (#1261)
* Ocpp21 smartcharging (#1096)

* Add smart charging fields.

Signed-off-by: Maaike Zijderveld <maaikez@users.noreply.github.com>
Signed-off-by: Martin Litre <mnlitre@gmail.com>
Signed-off-by: Piet Gömpel <pietgoempel@gmail.com>
Signed-off-by: Piet Gömpel <37657534+Pietfried@users.noreply.github.com>
Co-authored-by: Kai Hermann <kai-uwe.hermann@pionix.de>
Signed-off-by: Martin Litre <mnlitre@gmail.com>
Signed-off-by: Sebastian Lukas <sebastian.lukas@pionix.de>

* feat(OCPP2.1): Add AllowedEnergyTransferModes to auth validator (#1200)

* feat(OCPP2.1): Add AllowedEnergyTransferModes to auth validator

This list of allowed energy transfer modes has been added to the auth
validator interface so that the evse manager can be aware of what energy
transfer modes are allowed so that it can properly update the ISO15118
stack.

Signed-off-by: Martin Litre <mnlitre@gmail.com>

* feat: Add update energy transfer modes to both EvseManager/Iso15518

Added the possibility to update the supported energy transfer modes at
run time. This is necessary for use case Q01 for OCPP 2.1 where we want
to dynamically control whether or not we want to allow V2X.

Signed-off-by: Martin Litre <mnlitre@gmail.com>

---------

Signed-off-by: Martin Litre <mnlitre@gmail.com>

* feat(IsoExtension/EvseManager): Add vars to interfaces (#1224)

We need some information about the EV to fill out the ConnectedEV
component, this is now possible thanks to this new var of the
iso15118_extensions interface.
We also added information whether service renegotiation is supported.

A new config was added to the evse manager to list the supported
operation modes. We also added vars to transmit what is the full list of
supported energy transfers and operation modes.

Signed-off-by: Martin Litre <mnlitre@gmail.com>

* feat(OCPPv2.x): Add support for missing conversions for v2x params (#1254)

Added support for missing conversion of the notify ev charging needs
message.

Signed-off-by: Piet Gömpel <37657534+Pietfried@users.noreply.github.com>
Signed-off-by: Martin Litre <mnlitre@gmail.com>

---------

Signed-off-by: Maaike Zijderveld <maaikez@users.noreply.github.com>
Signed-off-by: Martin Litre <mnlitre@gmail.com>
Signed-off-by: Piet Gömpel <pietgoempel@gmail.com>
Signed-off-by: Piet Gömpel <37657534+Pietfried@users.noreply.github.com>
Signed-off-by: Sebastian Lukas <sebastian.lukas@pionix.de>
Co-authored-by: Maaike Zijderveld <maaikez@users.noreply.github.com>
Co-authored-by: Kai Hermann <kai-uwe.hermann@pionix.de>