fix: The valid_until part of EnforcedLimits is a date and time string
This identifies when a limit is no longer valid. The problem occurs when
there is a system clock update because the relationship between the current
time and valid_until is no longer correct.

If the clock jumps forward then the most recent EnforcedLimits can be considered
expired. If the clock jumps backwards the EnforcedLimits will be consider valid
for too long.

This change changes valid_until to valid_until_ts which is a duration in
seconds from the steady_clock (a monotonic clock that only ever increments).

Note: the unit tests fail because there isn't a simple mechanism to guess
what valid_until_ts will contain.

Signed-off-by: James Chapman <james.chapman@pionix.de>