fix: adjusted approach following feedback
The interface has been updated so that the number of seconds a limit is
valid for is specified rather than a time stamp.

This removes the dependency on system time especially where multiple Linux
instances exist.

As a result the unit tests now have an expected fixed value and now pass.

Signed-off-by: James Chapman <james.chapman@pionix.de>