EnergyManager to use steady_clock for valid until timeouts. (#1277)
* fix: The valid_until part of EnforcedLimits is a date and time string

This identifies when a limit is no longer valid. The problem occurs when
there is a system clock update because the relationship between the current
time and valid_until is no longer correct.

If the clock jumps forward then the most recent EnforcedLimits can be considered
expired. If the clock jumps backwards the EnforcedLimits will be consider valid
for too long.

This change changes valid_until to valid_until_ts which is a duration in
seconds from the steady_clock (a monotonic clock that only ever increments).

Note: the unit tests fail because there isn't a simple mechanism to guess
what valid_until_ts will contain.

Signed-off-by: James Chapman <james.chapman@pionix.de>

* fix: adjusted approach following feedback

The interface has been updated so that the number of seconds a limit is
valid for is specified rather than a time stamp.

This removes the dependency on system time especially where multiple Linux
instances exist.

As a result the unit tests now have an expected fixed value and now pass.

Signed-off-by: James Chapman <james.chapman@pionix.de>

* fix: PR 1277 fix - removed redundant using declarations
     consistent use of std::chrono::steady_clock vs steady_clock

Signed-off-by: James Chapman <james.chapman@pionix.de>

---------

Signed-off-by: James Chapman <james.chapman@pionix.de>