RpcApi: add a Python client GUI
This is a Python client for the JSON RPC API, written with the assistance of
AI.

It assists in manually exercizing the API, with automatic API.Hello on
connect, and pre-saved Method templates. It splits JSON RPC Method
responses and JSON RPC Notifications into separate pains, and allows
filtering the Notifications.

Signed-off-by: Moritz Barsnick <moritz.barsnick@chargebyte.com>