Allow defining Rust modules in different CMake scopes
While attempting to build an out-of-tree Rust module with CMake I noticed
that the `RUST_WORKSPACE_DIR` variable wasn't set for my module, which
broke the `rust_symlink_module_*` targets. I believe this happens because
CMake variables aren't consider in scope after calls to functions like
`add_subdirectory()`.

This commit ensures the variable is always in scope by storing it as a
target property, which is later retrieved whenever required. I went with
a target property instead of a cache variable to avoid potential naming
conflicts, seeing how those are not namespaced.

Signed-off-by: Ivar Scholten <ivar.scholten@protonmail.com>