Allow cross compiling Rust modules with CMake
The Rust integration for CMake runs Cargo (the Rust build system) under
the hood, but it previously didn't pass any information regarding the
target platform along. This resulted in Cargo always performing a build
for the host platform, which is incorrect in cross-compilation
environments.

This commit fixes that by the appropriate target to Cargo.

Signed-off-by: Ivar Scholten <ivar.scholten@protonmail.com>