feat(EvManager): Create a generic plugin command that uses a config key (#1518)
* feat(EvManager): Create a generic plugin command that uses a config key

We now can use a generic "plugin" command for the car simulation. This
avoids having to send a specific list of commands over mqtt. The list of
commands to run are now set in the "plugin_commands" configuration key.
This allows us to create an EVerest SIL configuration where the plugin
command can be set according to the EVSE type to avoid failed charging
session attempts because we are trying to start a basic charging session
in DC (for example).

Signed-off-by: Martin Litre <mnlitre@gmail.com>

* Setting the custom config cmd list via the new update_modify_charging_session_cmds()

Signed-off-by: Sebastian Lukas <sebastian.lukas@pionix.de>

---------

Signed-off-by: Martin Litre <mnlitre@gmail.com>
Signed-off-by: Sebastian Lukas <sebastian.lukas@pionix.de>
Co-authored-by: Sebastian Lukas <sebastian.lukas@pionix.de>