EvseV2G: fix randomness of SessionID (#1554)
* EvseV2G: fix logging of SessionIDs

They were printed in decimal, despite displaying a leading "0x". (It should
have used the format identifier PRIx64, not PRIu64.)

Also fix the fixed width prefix from 8 to 16 nibbles for the 64-bit
SessionID.

Also invert the byte order (consider it stored in big-endian). It now
displays like in Wireshark.

Signed-off-by: Moritz Barsnick <moritz.barsnick@chargebyte.com>

* EvseV2G: use correct amount of randomness for SessionID

Use 8 bytes for the 64-bit SessionID. This was an oversight in
https://github.com/EVerest/everest-core/pull/1407.

Also apply the modified generation to DIN, which was also missed in
https://github.com/EVerest/everest-core/pull/1407.

Signed-off-by: Moritz Barsnick <moritz.barsnick@chargebyte.com>

---------

Signed-off-by: Moritz Barsnick <moritz.barsnick@chargebyte.com>