feat(over_voltage): Internal over-voltage monitor implementation in EVSEManager
Implementation of an internal over-voltage monitor on top of the external one. This is a safety mechanism that reads the measured values received from over-voltage monitor and raises the over voltage faults (in case the external one is not reacting correctly).
There 2 limits that can be set: an error and an emergency limit.
In case of an voltage over the error limit, we are checking if this is just a very short peak or it will stay over 400ms (default value - configurable). If at the end of the 400ms we stayed above the error limit, we trigger the MREC5OverVoltage error.
In case of an voltage over the emergency error, we trigger ASAP the MREC5OverVoltage fault.Implementation of an internal over-voltage monitor on top of the external one. This is a safety mechanism that reads the measured values received from over-voltage monitor and raises the over voltage faults (in case the external one is not reacting correctly).
There 2 limits that can be set: an error and an emergency limit.
In case of an voltage over the error limit, we are checking if this is just a very short peak or it will stay over 400ms (default value - configurable). If at the end of the 400ms we stayed above the error limit, we trigger the MREC5OverVoltage error.
In case of an voltage over the emergency error, we trigger ASAP the MREC5OverVoltage fault.

Signed-off-by: Florin Mihut <florinmihut1@gmail.com>
Co-authored-by: Piet Gömpel <37657534+Pietfried@users.noreply.github.com>