EvseV2G, Evse15118D20, IsoMux, DummyV2G: Add new cmd update_supported_app_protocols and new var supported_app_protocols_secc
This commit adds a new command `update_supported_app_protocols` to the EvseV2G, Evse15118D20, IsoMux, and DummyV2G modules. This command allows updating the supported application protocols for the SECC (Supply Equipment Communication Controller). Additionally, a new variable `supported_app_protocols_secc` is introduced to store the current supported application protocols for the SECC.
Note: This command is intended for use with the Fake DC to disable the support of protocols in case of incompatibles between SECC and EVCC.

Signed-off-by: Fabian Hartung <fabian.hartung@chargebyte.com>